+++ R/package.R
@@ -1,10 +1,17 @@
 
-#' Fast sf-to-raster conversion
+#' @keywords internal
+#' @aliases fasterize-package
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL
+
+#' Fast polygons-to-raster conversion
 #'
-#' Fast sf-to-raster conversion
-#'
-#' @docType package
-#' @name fasterize
+#' Fast polygons-to-raster conversion
+#' @name fasterize-package
+
 NULL
 
 #' @useDynLib fasterize