+++ NEWS.md
@@ -1,4 +1,15 @@
-# fasterize dev
+# fasterize 1.0.5
+
+* Input geometries may now be of type "sfc_GEOMETRY" or other mixed types. Any non polygon
+ geometry is dropped with a message. 
+ 
+* 'field' may now be a numeric, integer, or factor vector or (as before) a name of a column in 
+  the input data frame. This means we can `fasterize(vec, r, field = seq_along(vec))` or
+  `fasterize(df, r, field = seq_len(dim(df)[1L]))` or with any values we like. 
+
+* Now supporting any geometry or dataframe input supported by wk (wkb, wkt, geos, as well as sf). 
+
+* Namespaced documentation fixes thanks to CRAN. 
 
 * Removed raster raster() method, not needed (raster handles sf now to obtain the 6 numbers and a string).